# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Taxomind LMS API
  description: |
    Comprehensive API documentation for Taxomind Learning Management System.
    
    ## Overview
    Taxomind is an intelligent LMS featuring AI-powered adaptive learning, real-time analytics, 
    and enterprise-grade security. The API supports multiple learning paths with role-based access 
    control for students, teachers, and administrators.
    
    ## Authentication
    All authenticated endpoints require a Bearer token in the Authorization header.
    Use the `/api/auth/login` endpoint to obtain a token.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse:
    - Anonymous: 10 requests per minute
    - Authenticated: 100 requests per minute
    - Premium: 1000 requests per minute
    
  version: 1.0.0
  contact:
    name: Taxomind Support
    email: support@taxomind.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://staging.taxomind.com
    description: Staging server
  - url: https://api.taxomind.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
        image:
          type: string
          format: uri
        emailVerified:
          type: string
          format: date-time
        isTwoFactorEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        imageUrl:
          type: string
          format: uri
        price:
          type: number
          format: float
          minimum: 0
        level:
          type: string
          enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        categoryId:
          type: string
        isPublished:
          type: boolean
        userId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        category:
          $ref: '#/components/schemas/Category'
        user:
          $ref: '#/components/schemas/User'
        _count:
          type: object
          properties:
            chapters:
              type: integer
            Enrollment:
              type: integer
            reviews:
              type: integer
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
        hasMore:
          type: boolean
paths:
...
  /api/courses:
    get:
      tags:
        - Courses
      summary: List all courses
      description: Get paginated list of published courses with optional filters
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: level
          in: query
          schema:
            type: string
            enum: [BEGINNER, INTERMEDIATE, ADVANCED]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: sort
          in: query
          schema:
            type: string
            enum: [title, price, createdAt, rating]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Course'

    post:
      tags:
        - Courses
      summary: Create a new course
      description: Create a new course (requires teacher role)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - description
                - categoryId
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 100
                description:
                  type: string
                  maxLength: 1000
                imageUrl:
                  type: string
                  format: uri
                price:
                  type: number
                  minimum: 0
                level:
                  type: string
                  enum: [BEGINNER, INTERMEDIATE, ADVANCED]
                categoryId:
                  type: string
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Course'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - requires teacher role
...
security:
  - bearerAuth: []